refactor backends: offscreen/window
