--- src/gfx/metal.rs
@@ -1,364 +1,0 @@
-#![allow(unexpected_cfgs)]
-use std::cell::RefCell;
-use std::sync::OnceLock;
-use std::time::{Instant, Duration};
-use metal::{
-    foreign_types::ForeignTypeRef,
-    CommandQueue, Device, DeviceRef, MTLDeviceLocation,
-};
-use skia_safe::{ImageInfo, Surface};
-use skia_safe::gpu::{
-    mtl, direct_contexts, surfaces, Budgeted, DirectContext, SurfaceOrigin
-};
-use objc::rc::autoreleasepool;
-use serde_json::{json, Value};
-
-use crate::context::page::ExportOptions;
-
-thread_local!( static MTL_CONTEXT: RefCell<Option<MetalContext>> = const { RefCell::new(None) }; );
-static MTL_CONTEXT_LIFESPAN:Duration = Duration::from_millis(2500); // rebuild is cheap, so use a short lifespan
-static MTL_STATUS: OnceLock<Value> = OnceLock::new();
-
-//
-// Offscreen rendering
-//
-pub struct MetalEngine {}
-
-impl MetalEngine {
-    pub fn api() -> Option<String>{
-        Some("Metal".to_string())
-    }
-
-    pub fn supported() -> bool {
-        Self::status()["renderer"] == "GPU"
-    }
-
-    pub fn status() -> Value {
-        MTL_STATUS.get_or_init(||{
-            // test whether a context can be created and do some one-time init if so
-            match MetalContext::new(){
-                Some(context) => {
-                    let device_name = format!("{} ({})", match context.device.location(){
-                        MTLDeviceLocation::BuiltIn => "Integrated GPU",
-                        MTLDeviceLocation::Slot => "Discrete GPU",
-                        MTLDeviceLocation::External => "External GPU",
-                        _ => "Other GPU"
-                    }, context.device.name());
-
-                    json!({
-                        "renderer": "GPU",
-                        "api": "Metal",
-                        "device": device_name,
-                        "threads": rayon::current_num_threads(),
-                    })
-                }
-                None => json!({
-                    "renderer": "CPU",
-                    "api": "Metal",
-                    "device": "CPU-based renderer (Fallback)",
-                    "threads": rayon::current_num_threads(),
-                    "error": "GPU initialization failed",
-                })
-            }
-        }).clone()
-    }
-
-    pub fn with_context<T, F>(f:F) -> Result<T, String>
-        where F:FnOnce(&mut MetalContext) -> Result<T, String>
-    {
-        match MetalEngine::supported() {
-            false => Err("Metal API not supported".to_string()),
-            true => MTL_CONTEXT.with_borrow_mut(|local_ctx|
-                autoreleasepool(||
-                    // lazily initialize this thread's context...
-                    local_ctx
-                        .take()
-                        .or_else(|| MetalContext::new() )
-                        .ok_or("Metal initialization failed".to_string())
-                        .and_then(|ctx|{
-                            f(local_ctx.insert(ctx))
-                        })
-                )
-            )
-        }
-    }
-
-    pub fn with_direct_context<F>(f:F)
-        where F:FnOnce(Option<&mut DirectContext>)
-    {
-        Self::with_context(|ctx| Ok(f(Some(&mut ctx.context))) ).ok();
-    }
-
-    pub fn make_surface(image_info: &ImageInfo, opts:&ExportOptions) -> Result<Surface, String>{
-        Self::with_context(|ctx| ctx.surface(image_info, opts) )
-    }
-
-    // allow the render thread to check how long the context has gone unused
-    pub fn context_is_idle() -> bool{
-        MTL_CONTEXT.with_borrow(|cell|
-            cell.as_ref().map(|engine| engine.last_use.elapsed() > MTL_CONTEXT_LIFESPAN).unwrap_or(false)
-        )
-    }
-
-    // called by the render thread when idle to drop the context & free gpu resources
-    pub fn evict_idle(){
-        MTL_CONTEXT.with_borrow_mut(|cell| {
-            cell.take_if(|engine| engine.last_use.elapsed() > MTL_CONTEXT_LIFESPAN);
-        });
-    }
-
-    // called by the render thread between jobs to expire skia's internal caches
-    pub fn purge_stale(){
-        MTL_CONTEXT.with_borrow_mut(|cell| {
-            if let Some(engine) = cell.as_mut(){
-                autoreleasepool(||
-                    engine.context.perform_deferred_cleanup(Duration::from_secs(1), None)
-                );
-            }
-        });
-    }
-
-    // create a job-specific autorelease pool for each pass through the render-thread
-    pub fn with_cleanup<T>(f: impl FnOnce() -> T) -> T {
-        autoreleasepool(f)
-    }
-}
-
-// create a Skia rendering context for use by either on- or offscreen renderers
-fn make_direct_context(device:&DeviceRef) -> Option<(CommandQueue, DirectContext)>{
-    let queue = device.new_command_queue();
-    let backend = unsafe {
-        mtl::BackendContext::new(
-            device.as_ptr() as mtl::Handle,
-            queue.as_ptr() as mtl::Handle,
-        )
-    };
-    direct_contexts::make_metal(&backend, None).map(|context| (queue, context))
-}
-
-pub struct MetalContext {
-    device: Device,
-    context: DirectContext,
-    msaa: Vec<usize>,
-    last_use: Instant,
-}
-
-impl MetalContext{
-    fn new() -> Option<Self>{
-        autoreleasepool(|| {
-            Device::system_default().and_then(|device|{
-                let last_use = Instant::now() + MTL_CONTEXT_LIFESPAN;
-                let msaa:Vec<usize> = [0,2,4,8,16,32].into_iter().filter(|s|{
-                    *s==0 || device.supports_texture_sample_count(*s as _)
-                }).collect();
-                make_direct_context(&device)
-                    .map(|(_queue, context)| MetalContext{device, context, msaa, last_use})
-            })
-        })
-    }
-
-    fn surface(&mut self, image_info: &ImageInfo, opts:&ExportOptions) -> Result<Surface, String> {
-        self.last_use = self.last_use.max(Instant::now());
-        surfaces::render_target(
-            &mut self.context,
-            Budgeted::Yes,
-            image_info,
-            Some(opts.msaa_from(&self.msaa)?),
-            SurfaceOrigin::BottomLeft,
-            Some(&opts.surface_props()),
-            false,
-            None
-        ).ok_or(
-            format!("Could not allocate new {}×{} bitmap", image_info.width(), image_info.height())
-        )
-    }
-
-}
-
-//
-// Windowed rendering
-//
-
-#[cfg(feature = "window")]
-use {
-    std::sync::Arc,
-    metal::{foreign_types::ForeignType, MTLPixelFormat, MetalLayer},
-    skia_safe::{scalar, ColorType, Size, Matrix, Color, Paint, BlendMode, SurfaceProps, canvas::SrcRectConstraint},
-    skia_safe::gpu::backend_render_targets,
-    raw_window_metal::Layer,
-    core_graphics_types::geometry::CGSize,
-    objc::{msg_send, sel, sel_impl, runtime::{self, Object}},
-    winit::{
-        dpi::PhysicalSize,
-        window::Window,
-        raw_window_handle::{RawWindowHandle, HasWindowHandle},
-        event_loop::ActiveEventLoop,
-    },
-    crate::context::page::Page,
-    super::{RenderOutcome, RenderCache, RenderState::Resizing},
-};
-
-#[cfg(feature = "window")]
-#[allow(non_upper_case_globals)]
-#[link(name = "QuartzCore", kind = "framework")]
-extern "C" {
-    static kCAGravityTopLeft: *mut Object;
-    static kCAGravityBottomLeft: *mut Object;
-}
-
-#[cfg(feature = "window")]
-pub struct MetalRenderer {
-    window: Arc<Window>,
-    backend: MetalBackend,
-    layer: MetalLayer,
-    cache: RenderCache,
-    last_page_id: usize, // previous frame's page id — render-loop history, not cache state
-}
-
-#[cfg(feature = "window")]
-impl MetalRenderer{
-    pub fn for_window(_event_loop: &ActiveEventLoop, window:Arc<Window>) -> Self {
-        let device = Device::system_default().expect("Metal device not found");
-
-        let raw_window = window
-            .window_handle()
-            .expect("Failed to retrieve a window handle")
-            .as_raw();
-
-        let raw_layer = match raw_window {
-            RawWindowHandle::AppKit(handle) => unsafe { Layer::from_ns_view(handle.ns_view) },
-            RawWindowHandle::UiKit(handle) => unsafe { Layer::from_ui_view(handle.ui_view) },
-            _ => panic!("Unsupported window handle type"),
-        };
-
-        let layer = unsafe{
-            let mtl_layer = MetalLayer::from_ptr(raw_layer.into_raw().as_ptr().cast());
-            let gravity = match msg_send![mtl_layer.as_ptr(), contentsAreFlipped] {
-                runtime::YES => kCAGravityBottomLeft,
-                _ => kCAGravityTopLeft,
-            };
-            let _: () = msg_send![mtl_layer.as_ptr(), setContentsGravity: gravity];
-            mtl_layer
-        };
-        layer.set_device(&device);
-        layer.set_pixel_format(MTLPixelFormat::BGRA8Unorm);
-        layer.set_presents_with_transaction(false);
-        layer.set_display_sync_enabled(true);
-        layer.set_opaque(false);
-        layer.set_framebuffer_only(false); // to enable blend modes
-
-        let draw_size = window.inner_size();
-        layer.set_drawable_size(CGSize::new(draw_size.width as f64, draw_size.height as f64));
-
-        let backend = MetalBackend::for_layer(&layer);
-        let cache = RenderCache::default();
-
-        Self{window, layer, backend, cache, last_page_id:0}
-    }
-
-    pub fn resize(&mut self, size: PhysicalSize<u32>) {
-        let cg_size = CGSize::new(size.width as f64, size.height as f64);
-        self.layer.set_drawable_size(cg_size);
-        self.cache.state = Resizing;
-    }
-
-    pub fn draw(&mut self, page:Page, matrix:Matrix, props:SurfaceProps, matte:Color){
-        let (clip, _) = matrix.map_rect(page.bounds);
-        let dpr = self.window.scale_factor() as f32;
-        let sync = self.cache.state == Resizing;
-        let take_snapshot = self.cache.wants_snapshot(&page, matte, dpr, self.last_page_id);
-
-        let outcome = self.backend.render_to_layer(&self.layer, &self.window, sync, take_snapshot, &props, |canvas| {
-            // fill the full surface (including any letterboxing) with the window’s background
-            // color, then lay the raster cache (if any) over the content area
-            canvas.clear(matte);
-            if let Some((image, src, dst)) = self.cache.validate(&page, matte, dpr, clip){
-                let mut paint = Paint::default();
-                paint.set_blend_mode(BlendMode::Src); // cached frame already includes the matte
-                canvas.draw_image_rect(image, Some((src, SrcRectConstraint::Strict)), dst, &paint);
-            }
-
-            // draw newly added vector layers
-            canvas.scale((dpr, dpr))
-                .clip_rect(clip, None, Some(true));
-            for pict in page.layers.iter().skip(self.cache.depth()){
-                canvas.draw_picture(pict, Some(&matrix), None);
-            }
-        });
-
-        // cache frame contents for use as background of next render pass
-        if let RenderOutcome::Rendered(image) = outcome{
-            self.cache.update(image, &page, matte, dpr, clip);
-            self.last_page_id = page.id;
-        }
-    }
-}
-
-#[cfg(feature = "window")]
-pub struct MetalBackend {
-    skia_ctx: DirectContext,
-    queue: CommandQueue,
-}
-
-#[cfg(feature = "window")]
-impl MetalBackend {
-    pub fn for_layer(layer:&MetalLayer) -> Self{
-        let (queue, skia_ctx) = make_direct_context(&layer.device())
-            .expect("Metal: Failed to create Skia direct context");
-        Self { skia_ctx, queue }
-    }
-
-    fn render_to_layer<F>(&mut self, layer:&MetalLayer, window:&Window, sync:bool, snapshot:bool, props:&SurfaceProps, f:F) -> RenderOutcome
-        where F:FnOnce(&skia_safe::Canvas)
-    {
-      autoreleasepool(||{
-        // if layer pool is exhausted, just drop the frame and try again next tick
-        let Some(drawable) = layer.next_drawable() else {
-            return RenderOutcome::Skipped;
-        };
-
-        let drawable_size = {
-            let size = layer.drawable_size();
-            Size::new(size.width as scalar, size.height as scalar)
-        };
-
-        let backend_render_target = unsafe {
-            let texture_info =
-                mtl::TextureInfo::new(drawable.texture().as_ptr() as mtl::Handle);
-            backend_render_targets::make_mtl(
-                (drawable_size.width as i32, drawable_size.height as i32),
-                &texture_info,
-            )
-        };
-
-        let mut surface = surfaces::wrap_backend_render_target(
-            &mut self.skia_ctx,
-            &backend_render_target,
-            SurfaceOrigin::TopLeft,
-            ColorType::BGRA8888,
-            None,
-            Some(props),
-        ).expect("MetalBackend: could not create render target");
-
-        // pass the suface's canvas to the user-provided callback
-        f(surface.canvas());
-
-        self.skia_ctx.flush_and_submit();
-        self.skia_ctx.perform_deferred_cleanup(Duration::from_secs(1), None);
-
-        window.pre_present_notify();
-        let command_buffer = self.queue.new_command_buffer();
-        command_buffer.present_drawable(drawable);
-        command_buffer.commit();
-
-        // during resizes, ensure drawing is complete before returning
-        if sync{ command_buffer.wait_until_completed(); }
-
-        // copy the frame contents (for the RenderCache) only when they'll be reused
-        RenderOutcome::Rendered(snapshot.then(||
-          surface.image_snapshot_with_bounds(surface.image_info().bounds())).flatten()
-        )
-      })
-    }
-
-}
